requirements.txt
pymongo==3.11.4
python-dotenv==0.17.1
Scrapy==2.5.0
